requirements.txt
# Telegram Bot
python-telegram-bot[asyncio]>=20.0

# Database and ORM
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.27.0
alembic>=1.11
psycopg2-binary>=2.9

# Redis and Queue
redis[asyncio]>=4.5.0
rq>=1.11

# HTTP and API calls
aiohttp>=3.8
httpx>=0.24

# Data validation and serialization
pydantic>=2.0

# Environment and configuration
python-dotenv>=1.0

# Logging (optional but recommended)
python-json-logger>=2.0

# Type hints and validation
typing-extensions>=4.5
